ingestion/requirements.txt
# Web & API
fastapi==0.104.1
uvicorn==0.24.0

supabase==2.11.0
httpx==0.27.2
kafka-python==2.0.2
pydantic==2.5.2
pydantic-settings==2.1.0

# Database Driver (if needed for direct SQL)
psycopg2-binary==2.9.9